wash the functions in each representations and the docstrings in representation
