+++ mrmustard/lab/representations/bargmann.py
@@ -15,7 +15,7 @@
 from mrmustard.math import Math
 from mrmustard.lab.representations.representation import Representation
 from mrmustard.lab.representations.data.qpoly_data import QPolyData
-from mrmustard.typing import Matrix, Scalar, Vector
+from mrmustard.typing import Matrix, Scalar, Vector, Tensor
 
 math = Math()
 
@@ -31,3 +31,38 @@ class Bargmann(Representation):
     def __init__(self, A: Matrix, b: Vector, c: Scalar) -> None:
         super().__init__()
         self.data = QPolyData(A=A, b=b, c=c)
+
+
+    @property
+    def purity(self) -> Scalar:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property    
+    def norm(self) -> float:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+
+
+    @property
+    def number_means(self) -> Vector:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property
+    def number_cov(self) -> Matrix:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property
+    def number_variances(self) -> int:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property
+    def number_stdev(self) -> int:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+
+
+    @property
+    def probability(self) -> Tensor:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
