+++ mrmustard/lab/representations/wavefunctionq.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import numpy as np
+from mrmustard.typing import Matrix, Scalar, Vector, Tensor
 from mrmustard.lab.representations.representation import Representation
 from mrmustard.lab.representations.data.wavefunctionarray_data import WavefunctionArrayData
 
@@ -26,3 +27,38 @@ class WaveFunctionQ(Representation):
 
     def __init__(self, qs:np.array, wavefunctionq:np.array) -> None:
         self.data = WavefunctionArrayData(qs=qs, array=wavefunctionq)
+    
+
+    @property
+    def purity(self) -> Scalar:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property    
+    def norm(self) -> float:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+
+
+    @property
+    def number_means(self) -> Vector:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property
+    def number_cov(self) -> Matrix:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property
+    def number_variances(self) -> int:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+    
+
+    @property
+    def number_stdev(self) -> int:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
+
+
+    @property
+    def probability(self) -> Tensor:
+        raise NotImplementedError(f"This property is not available in {self.__class__.__qualname__} representation")
